# Improve performance of pragma-region attach/detach by using transformers